* Changed userdata.sh to use the numbered scripts in cloud-init bin. * minor corrections to main README.
At the moment all scripts (except the one for enovation) are identical.

We need to discuss how we want each project to be able to make project-specific
changes (beyond using puppet) but without abusing the Packer or Makefile to suit
all edge cases.

** Also what's with all the userdata `service enovation restart` commands?**

I'm guessing that this is because chkconfig isn't on?
Is puppet managing the _ensure_ state of the service after deployment?

Or does the app only start correctly _the 2nd time_?

Surely this isn't needed for payment or booking?